+++ db/db_test_util.h
@@ -872,7 +872,11 @@ class CacheWrapper : public Cache {
     return target_->GetCharge(handle);
   }
 
-  void ApplyToAllCacheEntries(void (*callback)(void*, size_t),
+  DeleterFn GetDeleter(Handle* handle) const override {
+    return target_->GetDeleter(handle);
+  }
+
+  void ApplyToAllCacheEntries(void (*callback)(void* value, size_t charge),
                               bool thread_safe) override {
     target_->ApplyToAllCacheEntries(callback, thread_safe);
   }